GPIO callback added
Allow the user to provide their own function for setting/clearing a gpio
around actual event execution.  To be used to indicate utilization by
attaching to an LED or probing with a scope.  Compile switch is added so
you don't have to call any gpio function and can improve your code
efficiency.